Add solution for employee bonus problem (SQL)
Added an SQL script to create the Employee and Bonus tables and insert sample data. The script also includes a query to report the name and bonus amount of each employee with a bonus less than 1000. If an employee does not have a bonus, it will show as NULL.

Method used: Use a LEFT JOIN to combine the Employee and Bonus tables on the 'empId' column. Use the WHERE clause to filter out rows where the bonus is less than 1000 or is NULL.