Add solution for not boring movies problem (SQL)
Added an SQL script to create the Cinema table and insert sample data. The script also includes a query to report the movies with an odd-numbered ID and a description that is not "boring". The result table is ordered by 'rating' in descending order.

Method used: Use a SELECT statement to retrieve all columns from the Cinema table where the 'description' is not "boring" and the 'id' is odd. Use the ORDER BY clause to sort the result by 'rating' in descending order.