Add solution for list the products ordered in a period problem (SQL)
Added an SQL script to create the Products and Orders tables and insert sample data. The script also includes a query to get the names of products that have at least 100 units ordered in February 2020 and their total units.

Method used: Use a SELECT statement to join the Products and Orders tables on the 'product_id' column. Filter the results to only include orders placed in February 2020. Use the GROUP BY clause to group the results by 'product_name' and the HAVING clause to only include products with at least 100 units ordered.