Add solution for find customer referee problem (SQL)
Added an SQL script to create the Customer table and insert sample data. The script also includes a query to find the names of customers who are not referred by the customer with 'id' = 2. The result table is returned in any order.

Method used: Use a SELECT statement to retrieve the 'name' column from the Customer table. Use the WHERE clause to filter out customers whose 'referee_id' is 2 or is NULL.