Add solution for actors and directors who cooperated at least three times problem (SQL)
Added an SQL script to create the ActorDirector table and insert sample data. The script also includes a query to find all the pairs of 'actor_id' and 'director_id' where the actor has cooperated with the director at least three times. The result table is returned in any order.

Method used: Use a SELECT statement to retrieve 'actor_id' and 'director_id' from the ActorDirector table. Group the results by 'actor_id' and 'director_id'. Use the HAVING clause to filter groups with a count of timestamp greater than or equal to 3.