Add solution for big countries problem (SQL)
Added an SQL script to create the World table and insert sample data. The script also includes a query to find the name, population, and area of countries that are considered big. A country is classified as big if it has an area of at least three million square kilometers or a population of at least twenty-five million.

Method used: Use a SELECT statement to retrieve the 'name', 'population', and 'area' columns from the World table. Apply the WHERE clause to filter countries that have a population of at least 25,000,000 or an area of at least 3,000,000 square kilometers.