Add solution for customer who visited but did not make any transactions problem (SQL)
Added an SQL script to create the Visits and Transactions tables and insert sample data. The script also includes a query to find the IDs of users who visited without making any transactions and the number of such visits. The result table is returned in any order.

Method used: Use a SELECT statement to retrieve 'customer_id' and the count of visits without transactions (aliased as 'count_no_trans') from the Visits table. Perform a LEFT JOIN with the Transactions table on the 'visit_id' column. Use the WHERE clause to filter visits that have no corresponding transaction (t.'visit_id' IS NULL). Group the results by 'customer_id'.