Add solution for customer placing the largest number of orders problem (SQL)
Added an SQL script to create the Orders table and insert sample data. The script also includes a query to find the 'customer_number' for the customer who has placed the largest number of orders.

Method used: Use a SELECT statement to retrieve the 'customer_number' from the Orders table. Group the results by 'customer_number' and order them by the count of 'order_number' in descending order. Use the LIMIT clause to ensure only one result (the customer with the most orders) is returned.