Add solution for classes more than 5 students problem (SQL)
Added an SQL script to create the Courses table and insert sample data. The script also includes a query to find all the classes that have at least five students. The result table can be returned in any order.

Method used: Use a SELECT statement to retrieve the 'class' column from the Courses table. Group the results by 'class' and use the HAVING clause to filter groups with a count of student greater than or equal to 5.