Add solution for rising temperature problem (SQL)
Added an SQL script to create the Weather table and insert sample data. The script also includes a query to find all dates' IDs where the temperature is higher compared to the previous date (yesterday). The result table can be returned in any order.

Method used: Use a SELECT statement to retrieve the 'id' of dates where the temperature is higher than the previous date. Perform a JOIN operation on the Weather table to compare the temperatures of adjacent days. Use the DATEDIFF function to ensure the comparison is made between consecutive days.