Add solution for find users with valid e-mails problem (SQL)
Added an SQL script to create the Users table and insert sample data. The script also includes a query to find the users who have valid emails. A valid email has a prefix name that starts with a letter and may contain letters, digits, underscore '_', period '.', and/or dash '-'. The domain must be '@leetcode.com'.

Method used: Use a SELECT statement to retrieve all columns from the Users table. Apply the WHERE clause to filter emails that match the regex pattern for a valid email.