Add solution for capital gain and loss problem (SQL)
Added an SQL script to create the Stocks table and insert sample data. The script also includes a query to report the capital gain or loss for each stock by calculating the total gain or loss after buying and selling the stock one or many times. The result table can be returned in any order.

Method used: Use a SELECT statement to retrieve the 'stock_name' and calculate the capital gain or loss (aliased as 'capital_gain_loss'). Use the CASE statement within the SUM function to add the 'price' if the operation is "Sell" and subtract the 'price' if the operation is "Buy". Group the results by 'stock_name'.