Add solution for last person to fit in the bus problem (SQL)
Added an SQL script to create the Queue table and insert sample data. The script also includes a query to find the 'person_name' of the last person that can fit on the bus without exceeding the weight limit of 1000 kilograms. Each person boards the bus in the order specified by the turn column.

Method used: Use a subquery to calculate the running total weight (w) for each person as they board the bus, ordered by the 'turn' column. Use the SUM function with the OVER clause to get the cumulative weight. Filter the results in the outer query to only include those whose cumulative weight is less than or equal to 1000 kilograms. Order the results by the cumulative weight in descending order and limit the output to the last person who can fit on the bus.