Add solution for second highest salary problem (SQL)
Added an SQL script to create the Employee table and insert sample data. The script also includes a query to find the second highest distinct salary from the Employee table. If there is no second highest salary, it returns null.

Method used: Use a JOIN operation on the Employee table to compare salaries. The query finds the maximum salary from the Employee table where the salary is less than the highest salary. This effectively identifies the second highest distinct salary.